+++ api/src/certification/session/domain/usecases/get-attendance-sheet.js
@@ -1,4 +1,4 @@
-import { UserNotAuthorizedToAccessEntityError } from '../errors.js';
+import { UserNotAuthorizedToAccessEntityError } from '../../../../shared/domain/errors.js';
 
 const getAttendanceSheet = async function ({
   userId,
